security test for 6.1.7

+++ dbms/src/Flash/Coprocessor/TiDBTime.h
@@ -21,7 +21,6 @@
 
 #include <Common/MyTime.h>
 #include <Core/Types.h>
-#include <Storages/Transaction/TiDB.h>
 
 namespace DB
 {
@@ -36,7 +35,7 @@ public:
     }
     MyDateTime my_date_time;
     UInt8 time_type;
-    Int32 fsp;
+    Int8 fsp;
     UInt64 toChunkTime() const
     {
         UInt64 ret = 0;
